Implemented mainValidation, shows percentage of validation for a TPG on a given nb of seeds

+++ PendulumEmbeddedSTMProject/Core/Inc/Pendulum/TPGPrograms.h
@@ -1,51 +1,51 @@
 /**
  * File generated with GEGELATI v1.3.1
- * On the 2024-04-19 16:32:58
+ * On the 2024-04-30 11:04:22
  * With the CodeGen::ProgramGenerationEngine.
  */
 
 #ifndef C_TPGPrograms_H
 #define C_TPGPrograms_H
 
-int P0();
-int P1();
-int P2();
-int P3();
-int P4();
-int P5();
-int P6();
-int P7();
-int P8();
-int P9();
-int P10();
-int P11();
-int P12();
-int P13();
-int P14();
-int P15();
-int P16();
-int P17();
-int P18();
-int P19();
-int P20();
-int P21();
-int P22();
-int P23();
-int P24();
-int P25();
-int P26();
-int P27();
-int P28();
-int P29();
-int P30();
-int P31();
-int P32();
-int P33();
-int P34();
-int P35();
-int P36();
-int P37();
-int P38();
-int P39();
-int P40();
+double P0();
+double P1();
+double P2();
+double P3();
+double P4();
+double P5();
+double P6();
+double P7();
+double P8();
+double P9();
+double P10();
+double P11();
+double P12();
+double P13();
+double P14();
+double P15();
+double P16();
+double P17();
+double P18();
+double P19();
+double P20();
+double P21();
+double P22();
+double P23();
+double P24();
+double P25();
+double P26();
+double P27();
+double P28();
+double P29();
+double P30();
+double P31();
+double P32();
+double P33();
+double P34();
+double P35();
+double P36();
+double P37();
+double P38();
+double P39();
+double P40();
 #endif
